stop calling features geometries closes #132
