+++ src/api.js
@@ -2,9 +2,9 @@
 
 import mapboxgl from 'mapbox-gl';
 
-import Line from './geometries/line';
-import Point from './geometries/point';
-import Polygon from './geometries/polygon';
+import Line from './feature_types/line';
+import Point from './feature_types/point';
+import Polygon from './feature_types/polygon';
 
 export default class API extends mapboxgl.Control {
 
@@ -89,7 +89,7 @@ export default class API extends mapboxgl.Control {
   }
 
   /**
-   * get all draw geometries
+   * get all draw features
    * @returns {Object} a GeoJSON feature collection
    */
   getAll() {
@@ -148,7 +148,7 @@ export default class API extends mapboxgl.Control {
 
   /**
    * get selected feature collection of features
-   * @return {Object} a feature collection of geometries that are selected
+   * @return {Object} a feature collection of features that are selected
    */
   getSelected() {
     return this._store.getSelectedIds()
@@ -173,7 +173,7 @@ export default class API extends mapboxgl.Control {
   }
 
   /**
-   * remove all geometries
+   * remove all features
    * @returns {Draw} this
    */
   clear() {
