+++ src/feature_types/polygon.js
@@ -1,49 +1,56 @@
 'use strict';
 
-import Geometry from './geometry';
+import Feature from './feature';
 import { translatePoint, DOM } from '../util';
 
 /**
- * Line geometry class
+ * Polygon geometry class
  *
  * @param {Object} options
- * @param {Map} options.map - Instance of MapboxGL Map
+ * @param {Map} options.map - Instance of MapboxGl Map
  * @param {Object} [options.data] - GeoJSON feature
- * @returns {Line} this
+ * @returns {Polygon} this
  */
-export default class Line extends Geometry {
+export default class Polygon extends Feature {
 
   constructor(options) {
     if (!options.data) {
       options.data = {
         geometry: {
-          coordinates: [[0, 0], [0, 0]]
+          coordinates: [[[0, 0],[0, 0], [0, 0], [0, 0]]]
         }
       };
     }
-    options.type = 'LineString';
+    options.type = 'Polygon';
     super(options);
 
-    // event listeners
+    // event handlers
     this.onMouseMove = this._onMouseMove.bind(this);
   }
 
   onClick(e) {
     var p = [ e.lngLat.lng, e.lngLat.lat ];
+
     if (typeof this.vertexIdx === 'undefined') {
-      this.coordinates = [p];
       this.vertexIdx = 0;
+      this.first = p;
+      this.coordinates = [[p, p, p, p]];
       this.ready = true;
-    } else {
-      this.coordinates.splice(-1, 1, p, p);
     }
+
+    if (this.vertexIdx > 1) {
+      this.coordinates[0].push(this.first);
+    }
+
+    this.coordinates[0][this.vertexIdx] = p;
+
     this.vertexIdx++;
   }
 
   _onMouseMove(e) {
     var pos = DOM.mousePos(e.originalEvent, this._map._container);
     var coords = this._map.unproject([pos.x, pos.y]);
-    this.coordinates[this.vertexIdx] = [ coords.lng, coords.lat ];
+    this.coordinates[0][this.vertexIdx] = [ coords.lng, coords.lat ];
   }
 
   onStopDrawing(e) {
@@ -52,8 +59,11 @@ export default class Line extends Geometry {
 
   onDoubleClick(e) {
     const ENTER = 13;
-    var subtract = e.keyCode === ENTER ? 1 : 2;
-    this.coordinates = this.coordinates.slice(0, this.coordinates.length - subtract);
+    if (this.vertexIdx > 2) {
+      var idx = this.vertexIdx - (e.keyCode === ENTER ? 0 : 1);
+      var remove = e.keyCode === ENTER ? 1 : 2;
+      this.coordinates[0].splice(idx, remove);
+    }
     super.onStopDrawing(e);
   }
 
@@ -63,18 +73,22 @@ export default class Line extends Geometry {
    * @param {Array<Number>} init - the position of the mouse at the start of the drag
    * @param {Array<Number>} curr - the current position of the mouse
    * @param {Number} idx - the index of the point being updated in `feature.geometry.coordinates`
+   * @private
    */
   moveVertex(init, curr, idx) {
     if (!this.movingVertex) {
       this.movingVertex = true;
-      this.initCoords = this.coordinates[idx];
+      this.initCoords = JSON.parse(JSON.stringify(this.coordinates[0][idx]));
     }
 
     var dx = curr.x - init.x;
     var dy = curr.y - init.y;
-    var newPoint = translatePoint(JSON.parse(JSON.stringify(this.initCoords)), dx, dy, this._map);
+    var newPoint = translatePoint(this.initCoords, dx, dy, this._map);
 
-    this.coordinates[idx] = newPoint;
+    this.coordinates[0][idx] = newPoint;
+    if (idx === 0) {
+      this.coordinates[0][this.coordinates[0].length - 1] = newPoint;
+    }
   }
 
   /**
@@ -82,9 +96,11 @@ export default class Line extends Geometry {
    *
    * @param {Object} coords - The coordinates of the new vertex in the for { lng: <Number>, lat: <Number> }
    * @param {Number} idx - the index at which the new point will be placed in `feature.geometry.coordinates`
+   * @private
    */
   addVertex(coords, idx) {
     coords = this._map.unproject(coords);
-    this.coordinates.splice(idx, 0, [ coords.lng, coords.lat ]);
+    this.coordinates[0].splice(idx, 0, [ coords.lng, coords.lat ]);
   }
+
 }
