+++ src/feature_types/square.js
